refactor(mobile): use shared diff and OSC implementations
