+++ mobile/src/session/mobile-diff-comments.ts
@@ -1,4 +1,7 @@
 import type { DiffComment, DiffReviewScope } from '../../../src/shared/types'
+import { formatDiffComment, formatDiffComments } from '../../../src/shared/diff-comments-format'
+
+export { formatDiffComment, formatDiffComments }
 
 export type CreateMobileDiffCommentInput = {
   worktreeId: string
@@ -16,42 +19,8 @@ function isRecord(value: unknown): value is Record<string, unknown> {
   return typeof value === 'object' && value !== null
 }
 
-function isMarkdownComment(comment: Pick<DiffComment, 'source'>): boolean {
-  return comment.source === 'markdown'
-}
-
 function normalizeScope(value: unknown): DiffReviewScope | undefined {
   return value === 'unstaged' || value === 'staged' || value === 'branch' ? value : undefined
-}
-
-// Why: mobile Vitest/Metro run from the mobile package and cannot transform
-// runtime imports from root src/shared. Keep this byte-for-byte compatible with
-// the desktop shared formatter contract.
-export function formatDiffComment(c: DiffComment): string {
-  const escaped = c.body
-    .replace(/\\/g, '\\\\')
-    .replace(/"/g, '\\"')
-    .replace(/\r/g, '\\r')
-    .replace(/\n/g, '\\n')
-  const locationLabel =
-    c.lineNumber === 0
-      ? 'Scope: file'
-      : c.startLine !== undefined && c.startLine !== c.lineNumber
-        ? `Lines: ${c.startLine}-${c.lineNumber}`
-        : `Line: ${c.lineNumber}`
-  if (!isMarkdownComment(c)) {
-    return [`File: ${c.filePath}`, locationLabel, `User comment: "${escaped}"`].join('\n')
-  }
-  return [
-    `File: ${c.filePath}`,
-    'Source: markdown',
-    locationLabel,
-    `User comment: "${escaped}"`
-  ].join('\n')
-}
-
-export function formatDiffComments(comments: readonly DiffComment[]): string {
-  return comments.map(formatDiffComment).join('\n\n')
 }
 
 export function formatMobileDiffReviewPrompt(comments: readonly DiffComment[]): string {
