--- mobile/src/terminal/terminal-osc-link-ranges.ts
@@ -1,18 +1,0 @@
-import type { TerminalOscLinkRange } from '../../../src/shared/terminal-osc-link-ranges'
-
-export type { TerminalOscLinkRange }
-
-export function isTerminalOscLinkRanges(value: unknown): value is TerminalOscLinkRange[] {
-  return (
-    Array.isArray(value) &&
-    value.every(
-      (entry) =>
-        entry != null &&
-        typeof entry === 'object' &&
-        Number.isInteger((entry as TerminalOscLinkRange).row) &&
-        Number.isInteger((entry as TerminalOscLinkRange).startCol) &&
-        Number.isInteger((entry as TerminalOscLinkRange).endCol) &&
-        typeof (entry as TerminalOscLinkRange).uri === 'string'
-    )
-  )
-}
